correction des conflits et rétablissement
