--- qsequoia2/scripts/layout_designer/__init__.py
@@ -1,2 +1,0 @@
-# dialogs/__init__.py
-# This file is required to make the dialogs folder a Python package.
